docs: sendonion created_at is u32
sendonion calls send_payment_core(), which reports via
json_sendpay_in_progress() -> json_add_payment_fields().
That function emits created_at with json_add_u32 from the u32
wallet_payment struct member:

  lightningd/pay.c:158
    json_add_u32(response, "created_at", t->timestamp);
  lightningd/pay.c:301
    json_add_payment_fields(response, payment);
  lightningd/pay.c:1384
    return send_payment_core(ld, cmd, ...);
  wallet/wallet.h:363
    u32 timestamp;

Changelog-None

+++ cln-rpc/src/model.rs
@@ -6968,7 +6968,7 @@ pub mod responses {
 	    // Path `SendOnion.status`
 	    pub status: SendonionStatus,
 	    pub amount_sent_msat: Amount,
-	    pub created_at: u64,
+	    pub created_at: u32,
 	    pub created_index: u64,
 	    pub id: u64,
 	    pub payment_hash: Sha256,
